making base code

+++ src/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id        int    `json:"_id"`
+	Id        int64  `json:"_id"`
 	Email     string `json:"email"`
 	Name      string `json:"name"`
 	Password  string `json:"password,omitempty"`
